custom chain, added config: maxBlockRange
