+++ ponder.address.ts
@@ -1,19 +1,8 @@
 import { Address, zeroAddress } from 'viem';
 import { Chain, mainnet } from 'viem/chains';
+import { ethereum3 } from './ponder.chains';
 
 export type AddressObject = { [chainId in Chain['id']]?: { [key: string]: Address } };
-
-export const ethereum3 = {
-	id: 1337,
-	name: 'Ethereum3',
-	nativeCurrency: { name: 'Ethereum3', symbol: 'ETH3', decimals: 18 },
-	rpcUrls: {
-		default: { http: ['https://ethereum3.3dotshub.com'] },
-	},
-	blockExplorers: {
-		default: { name: 'Blockscout', url: '' },
-	},
-} as const satisfies Chain;
 
 export const ADDRESS: AddressObject = {
 	[mainnet.id]: {
@@ -25,11 +14,11 @@ export const ADDRESS: AddressObject = {
 		wFPS: '0x5052D3Cc819f53116641e89b96Ff4cD1EE80B182',
 	},
 	[ethereum3.id]: {
-		frankenCoin: '0x7c0E7381D07e132747B425972a73F6045a26AB07',
+		frankenCoin: '0x9800f06718bB6F7FCAC181ED26753E2E670cb9e0',
 		bridge: zeroAddress,
 		xchf: zeroAddress,
-		equity: '0x8fcfC3a671809D104d9c50cC67eA8C5d7cC658E8',
-		mintingHub: '0x4800b6c288e4B2BBa7b2314328DB485F5FfB0414',
+		equity: '0x97e3bbF39138B1e7E1d06dd26E7E3b9d558b00b2',
+		mintingHub: '0xA0d6ce30a8a4eab09eD74f434dcA4Ce4169aDd03',
 		wFPS: zeroAddress,
 		// mockWbtc: '0x1b01c6b10ca8AeD4F1e0d39319aa27183BBC1578',
 		// mockLseth: '0xd54Fb4EE40ca7F0FeF1cd87AC81dE3F247776209',
